error.hpp: move cusparse error to seperate header.
